+++ model_components/temporal_compressor/conv1d_block_no_dilation_with_skip.py
@@ -1,3 +1,4 @@
+import torch
 from torch import nn
 from model_components.temporal_compressor.conv1d_block_no_dilation_no_skip import Conv1DBlockNoDilationNoSkip
 
@@ -34,7 +35,7 @@ class Conv1DBlockNoDilationWithSkip(nn.Module):
         self.skip_halving_conv = nn.Conv1d(
             in_channels=n_input_channels,
             out_channels=n_filters_skip,
-            kernel_size=2, stride=2, dtype=torch.float16)
+            kernel_size=2, stride=2, dtype=torch.float32)
 
         self.block = Conv1DBlockNoDilationNoSkip(
             block_num=block_num, input_len=input_len,
